+++ public/javascripts/backbone.eventbinder.js
@@ -1,3 +1,7 @@
+// Backbone.EventBinder, v0.1.0
+// Copyright (c)2012 Derick Bailey, Muted Solutions, LLC.
+// Distributed under MIT license
+// http://github.com/marionettejs/backbone.eventbinder
 // EventBinder
 // -----------
 //
@@ -17,7 +21,7 @@ Backbone.EventBinder = (function(Backbone, _){
   var handlerMap = {
     // 'default' type accounts for Backbone style objects extending
     // Backbone.Events
-    default : {
+    "default" : {
       bindTo : function (obj, eventName, callback, context) {
         context = context || this;
         obj.on(eventName, callback, context);
@@ -64,9 +68,9 @@ Backbone.EventBinder = (function(Backbone, _){
   // Use whatever best logic necessary to determine the type
   // of the supplied object
   function getHandlerForObject(obj) {
-    if (obj.jquery) return handlerMap.jquery;
+    if (obj.jquery) { return handlerMap.jquery; }
 
-    return handlerMap.default;
+    return handlerMap["default"];
   }
   
   // Constructor function