+++ changelog.md
@@ -1,5 +1,8 @@
 ### v1.0.0-beta2 [view commit logs](https://github.com/marionettejs/backbone.marionette/compare/v1.0.0-beta1...dev)
 
+* templateHelpers
+  * **BREAKING:** Changed when the templateHelpers is mixed in to the data for a view, so that it is no longer dependent on the `serializeData` implementation
+
 * CollectionView
   * Fixed bug where adding an item to a collection would not allow the CollectionView to propagate the itemView's events
   * Allow `itemViewOptions` to be specified in CollectionView constructor options